#437: Change 'state' to 'region' and 'zip' to 'postal code'
The desire is to move from supporting exclusively US mailing addresses to supporting international addresses, starting with Canada. Countries are not consistent with how they name their administrative divisions. "Region" appears to be a conventional term for the principal subdivisions of a country - whether they're states, provinces, or other.

Additionally, a ZIP code is a particular type of postal code, but the terminology "ZIP code" is specific to the US, so this changes to use a more generic term.